backend/requirements.txt
# FastAPI core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0

# Config & validation
pydantic>=2.7.1
pydantic-settings>=2.2.1
email-validator>=2.2.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
neo4j>=5.20.0
redis>=5.0.4

# AI / ML
openai>=1.30.1
spacy>=3.7.2
numpy>=1.24

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.6
pytest-mock>=3.14.0
anyio>=4.3.0

# Utilities
python-dotenv>=1.0.1
arq>=0.26.0
tenacity>=8.3.0
pypdf>=4.2.0